Refactor npm package info lookup
