+++ extensions/npm/src/features/packageInfo.ts
@@ -1,0 +1,187 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { MarkdownString, Uri, workspace, l10n } from 'vscode';
+import { XHRRequest } from 'request-light';
+
+import type * as cp from 'child_process';
+import { dirname } from 'path';
+import { fromNow } from './date';
+
+const USER_AGENT = 'Visual Studio Code';
+
+export interface ViewPackageInfo {
+	description: string;
+	version?: string;
+	time?: string;
+	homepage?: string;
+	installedVersion?: string;
+}
+
+/**
+ * Resolves package metadata for npm package names and formats hover content.
+ */
+export class NpmPackageInfoProvider {
+
+	public constructor(
+		private readonly xhr: XHRRequest,
+		private readonly npmCommandPath: string | undefined,
+	) {
+	}
+
+	public isEnabled(): boolean {
+		return !!this.npmCommandPath || this.onlineEnabled();
+	}
+
+	public getDocumentation(description: string | undefined, version: string | undefined, time: string | undefined, homepage: string | undefined, installedVersion: string | undefined): MarkdownString {
+		const str = new MarkdownString();
+		if (description) {
+			str.appendText(description);
+		}
+		if (version) {
+			str.appendText('\n\n');
+			str.appendText(time ? l10n.t("Latest version: {0} published {1}", version, fromNow(Date.parse(time), true, true)) : l10n.t("Latest version: {0}", version));
+		}
+		if (installedVersion) {
+			str.appendText('\n\n');
+			str.appendText(l10n.t("Installed version: {0}", installedVersion));
+		}
+		if (homepage) {
+			str.appendText('\n\n');
+			str.appendText(homepage);
+		}
+		return str;
+	}
+
+	public async fetchPackageInfo(pack: string, resource: Uri | undefined): Promise<ViewPackageInfo | undefined> {
+		if (!this.isValidNPMName(pack)) {
+			return undefined;
+		}
+
+		let info: ViewPackageInfo | undefined;
+		let installedVersion: string | undefined;
+		if (this.npmCommandPath) {
+			([info, installedVersion] = await Promise.all([
+				this.npmView(this.npmCommandPath, pack, resource),
+				this.npmListInstalledVersion(this.npmCommandPath, pack, resource)
+			]));
+		}
+		if (!info && this.onlineEnabled()) {
+			info = await this.npmjsView(pack);
+		}
+		if (installedVersion) {
+			info = info ?? { description: '' };
+			info.installedVersion = installedVersion;
+		}
+		return info;
+	}
+
+	private onlineEnabled(): boolean {
+		return !!workspace.getConfiguration('npm').get('fetchOnlinePackageInfo');
+	}
+
+	private isValidNPMName(name: string): boolean {
+		// following rules from https://github.com/npm/validate-npm-package-name,
+		// leading slash added as additional security measure
+		if (!name || name.length > 214 || name.match(/^[-_.\s]/)) {
+			return false;
+		}
+		const match = name.match(/^(?:@([^/~\s)('!*]+?)[/])?([^/~)('!*\s]+?)$/);
+		if (match) {
+			const scope = match[1];
+			if (scope && encodeURIComponent(scope) !== scope) {
+				return false;
+			}
+			const value = match[2];
+			return encodeURIComponent(value) === value;
+		}
+		return false;
+	}
+
+	private async npmView(npmCommandPath: string, pack: string, resource: Uri | undefined): Promise<ViewPackageInfo | undefined> {
+		// Request @latest to avoid fetching publish timestamps for all versions in the time field.
+		const args = ['view', '--json', '--', `${pack}@latest`, 'description', 'homepage', 'version', 'time'];
+
+		const stdout = await this.runNpmCommand(npmCommandPath, args, resource);
+		if (stdout) {
+			try {
+				const content = JSON.parse(stdout);
+				const version = content['version'];
+				return {
+					description: content['description'],
+					version: content['version'],
+					time: version ? content['time']?.[version] : undefined,
+					homepage: content['homepage']
+				};
+			} catch (e) {
+				// ignore
+			}
+		}
+		return undefined;
+	}
+
+	private async npmjsView(pack: string): Promise<ViewPackageInfo | undefined> {
+		const queryUrl = 'https://registry.npmjs.org/' + encodeURIComponent(pack);
+		try {
+			const success = await this.xhr({
+				url: queryUrl,
+				headers: { agent: USER_AGENT }
+			});
+			const obj = JSON.parse(success.responseText);
+			const version = obj['dist-tags']?.latest || Object.keys(obj.versions).pop() || '';
+			return {
+				description: obj.description || '',
+				version,
+				time: obj.time?.[version],
+				homepage: obj.homepage || ''
+			};
+		}
+		catch (e) {
+			// ignore
+		}
+		return undefined;
+	}
+
+	/**
+	 * Runs an npm command and returns its stdout, or undefined if the command fails.
+	 * Sets up cwd from resource, applies corepack env vars, and handles win32 shell quoting.
+	 */
+	private async runNpmCommand(npmCommandPath: string, args: string[], resource: Uri | undefined): Promise<string | undefined> {
+		const cp = await import('child_process');
+		return new Promise((resolve, _reject) => {
+			const cwd = resource && resource.scheme === 'file' ? dirname(resource.fsPath) : undefined;
+
+			// corepack npm wrapper would automatically update package.json. disable that behavior.
+			// COREPACK_ENABLE_AUTO_PIN disables the package.json overwrite, and
+			// COREPACK_ENABLE_PROJECT_SPEC makes the npm view command succeed
+			//   even if packageManager specified a package manager other than npm.
+			const env = { ...process.env, COREPACK_ENABLE_AUTO_PIN: '0', COREPACK_ENABLE_PROJECT_SPEC: '0' };
+			let options: cp.ExecFileOptions = { cwd, env };
+			let commandPath: string = npmCommandPath;
+			if (process.platform === 'win32') {
+				options = { cwd, env, shell: true };
+				commandPath = `"${npmCommandPath}"`;
+			}
+			cp.execFile(commandPath, args, options, (error, stdout) => {
+				resolve(error ? undefined : stdout);
+			});
+		});
+	}
+
+	private async npmListInstalledVersion(npmCommandPath: string, pack: string, resource: Uri | undefined): Promise<string | undefined> {
+		const args = ['ls', '--json', '--depth=0', '--', pack];
+		const stdout = await this.runNpmCommand(npmCommandPath, args, resource);
+		if (stdout) {
+			try {
+				const content = JSON.parse(stdout);
+				const version = content?.dependencies?.[pack]?.version;
+				return typeof version === 'string' ? version : undefined;
+			} catch (e) {
+				// ignore
+			}
+		}
+		return undefined;
+	}
+}
